refactor: ToolResultsPanel → SessionSidebar (#842)
The component was named after its primary content (the tool-result
preview list) but had drifted into being the whole left sidebar:
role header, layout / right-sidebar toggles, the result list, AND
the running indicator. The testid `[sidebar-role-header]` already
admits the truth — only the component name was lagging.

Changes (Approach A from the issue: rename only):
- src/components/ToolResultsPanel.vue → SessionSidebar.vue
  (`git mv`, no internal changes; tab-bar / testid layout intact)
- src/App.vue: import + tag + ref renamed (toolResultsPanelRef →
  sessionSidebarRef)
- src/composables/useChatScroll.ts: parameter renamed to match
- src/utils/agent/formatElapsed.ts, src/components/SessionHistoryPanel.vue,
  e2e/tests/streaming-autoscroll.spec.ts: comment text updated
- test/composables/test_useChatScroll.ts: parameter usage updated
- docs/ui-cheatsheet.md: new <SessionSidebar> ASCII block per the
  same-PR maintenance discipline introduced in #841

Out of scope (per #842):
- No internal split of the component into header / list / running
  indicator. (B) was considered and not done — the secondary
  responsibilities are tiny and an internal split would just add
  indirection. Re-evaluate if SessionSidebar grows further.
- No testid changes — `[sidebar-role-header]` and
  `[tool-results-scroll]` stay as-is so e2e is unaffected.
- packages/mulmoclaude/src/ vendored copy stays untouched —
  prepare-dist.js regenerates it at the next /publish-mulmoclaude.
- plans/done/* historical references keep the old name.

Refs: #842
