+++ src/App.vue
@@ -84,9 +84,9 @@
 
       <!-- Sidebar (Single layout only) -->
       <div v-if="!isStackLayout && !sidePanelExpanded" class="w-80 flex-shrink-0 border-r border-gray-200 flex flex-col bg-white text-gray-900 relative">
-        <!-- Tool result previews -->
-        <ToolResultsPanel
-          ref="toolResultsPanelRef"
+        <!-- Tool result previews + role header + running indicator (#842) -->
+        <SessionSidebar
+          ref="sessionSidebarRef"
           :results="sidebarResults"
           :selected-uuid="selectedResultUuid"
           :result-timestamps="activeSession?.resultTimestamps ?? new Map()"
@@ -221,7 +221,7 @@ import SuggestionsPanel from "./components/SuggestionsPanel.vue";
 import ChatInput, { type PastedFile } from "./components/ChatInput.vue";
 import SessionHistoryExpandButton from "./components/SessionHistoryExpandButton.vue";
 import SessionHistoryPanel from "./components/SessionHistoryPanel.vue";
-import ToolResultsPanel from "./components/ToolResultsPanel.vue";
+import SessionSidebar from "./components/SessionSidebar.vue";
 import PluginLauncher from "./components/PluginLauncher.vue";
 import StackView from "./components/StackView.vue";
 import FilesView from "./components/FilesView.vue";
@@ -430,11 +430,11 @@ const { mergedSessions, tabSessions } = useMergedSessions({
 // /api/sessions refetch.
 useFaviconState({ isRunning, sessions: mergedSessions, sessionsUnreadCount: unreadCount, cpuLoadRatio });
 
-const toolResultsPanelRef = ref<{ root: HTMLDivElement | null } | null>(null);
+const sessionSidebarRef = ref<{ root: HTMLDivElement | null } | null>(null);
 const canvasRef = ref<HTMLDivElement | null>(null);
 const chatInputRef = ref<{ focus: () => void } | null>(null);
 const { focusChatInput } = useChatScroll({
-  toolResultsPanelRef,
+  sessionSidebarRef,
   toolResults,
   isRunning: activeSessionRunning,
   chatInputRef,
