+++ src/components/SessionSidebar.vue
